systemd socket activation compatibility
Runc already supports systemd socket activation (a.k.a. it detects LISTEN_PID and LISTEN_FDS), but the containerID validation is too strict, which causes a problem when runc is used with systemd socket Accept=true (a.k.a. container-per-connection) style. This tiny resolves that problem and makes it possible to start a new container dynamically for every new socket.