Systemd socket activation compatibility
Runc already supports systemd socket activation (a.k.a. it detects
LISTEN_PID and LISTEN_FDS), but the containerID validation was too
strict. This caused a problem when runc is used with systemd
container-per-connection style (a.k.a. Accept=true)

This tiny change makes it possible to start a new container for every
new socket dynamically.